fix: break 37-file MCP cycle by extracting McpToolContext to mcp/types.ts
McpToolContext was defined in server.ts, which imported TOOL_HANDLERS from
tools/index.ts (the barrel). Every tool module imported McpToolContext back
from server.ts, creating a 37-file circular dependency flagged by codegraph
cycles in two consecutive architectural audits.

Fix: extract McpToolContext and McpToolHandler into src/mcp/types.ts, which
only depends on db/index.js (outside the MCP subtree). server.ts and all 35
tool modules now import from types.ts instead of server.ts, eliminating the
cycle. server.ts re-exports McpToolContext for backward compatibility.
