+++ src/mcp/tools/code-owners.ts
@@ -1,4 +1,4 @@
-import type { McpToolContext } from '../server.js';
+import type { McpToolContext } from '../types.js';
 
 export const name = 'code_owners';
