+++ src/mcp/tools/dataflow.ts
@@ -1,5 +1,5 @@
 import { effectiveOffset, MCP_DEFAULTS } from '../middleware.js';
-import type { McpToolContext } from '../server.js';
+import type { McpToolContext } from '../types.js';
 
 export const name = 'dataflow';
