+++ src/mcp/tools/triage.ts
@@ -1,6 +1,6 @@
 import type { Role } from '../../types.js';
 import { effectiveLimit, effectiveOffset, MCP_DEFAULTS, MCP_MAX_LIMIT } from '../middleware.js';
-import type { McpToolContext } from '../server.js';
+import type { McpToolContext } from '../types.js';
 
 export const name = 'triage';
